requirements.txt
Django==2.2
django-environ==0.4.5
django-cors-headers==3.0
django-storages==1.7.1
djangorestframework==3.9
Pillow==6.0
psycopg2-binary==2.8
pytz==2019.1
